+++ app/src/androidTest/java/org/mozilla/focus/activity/AdBlockingTest.java
@@ -17,11 +17,12 @@ import android.support.test.uiautomator.UiSelector;
 import org.junit.After;
 import org.junit.Rule;
 import org.junit.Test;
+import org.mozilla.focus.helpers.TestHelper;
 
 import static android.support.test.espresso.action.ViewActions.click;
 import static junit.framework.Assert.assertTrue;
 import static junit.framework.TestCase.assertFalse;
-import static org.mozilla.focus.activity.TestHelper.waitingTime;
+import static org.mozilla.focus.helpers.TestHelper.waitingTime;
 import static org.mozilla.focus.fragment.FirstrunFragment.FIRSTRUN_PREF;
 
 // This test checks basic ad blocking capability
