+++ app/src/androidTest/java/org/mozilla/focus/activity/screenshots/HomeScreenScreenshots.java
@@ -13,7 +13,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mozilla.focus.R;
 import org.mozilla.focus.activity.MainActivity;
-import org.mozilla.focus.activity.TestHelper;
+import org.mozilla.focus.helpers.TestHelper;
 import org.mozilla.focus.helpers.MainActivityFirstrunTestRule;
 
 import tools.fastlane.screengrab.Screengrab;
