+++ app/src/androidTest/java/org/mozilla/focus/helpers/TestHelper.java
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.focus.activity;
+package org.mozilla.focus.helpers;
 
 import android.os.RemoteException;
 import android.support.annotation.NonNull;
